move kickout code to slasher
