+++ action/protocol/rewarding/protocol_test.go
@@ -148,25 +148,33 @@ func testProtocol(t *testing.T, test func(*testing.T, context.Context, protocol.
 	}
 	cfg := config.Default
 	committee := mock_committee.NewMockCommittee(ctrl)
-	pp, err := poll.NewGovernanceChainCommitteeProtocol(
-		nil,
+	slasher, err := poll.NewSlasher(
+		&cfg.Genesis,
+		func(context.Context, uint64) (uint64, map[string]uint64, error) {
+			return 0, nil, nil
+		},
 		func(protocol.StateReader, uint64) ([]*state.Candidate, error) {
 			return abps, nil
 		},
 		nil,
 		nil,
 		nil,
+		nil,
+		5,
+		5,
+		cfg.Genesis.ProductivityThreshold,
+		cfg.Genesis.KickoutEpochPeriod,
+		cfg.Genesis.UnproductiveDelegateMaxCacheSize,
+		cfg.Genesis.KickoutIntensityRate)
+	require.NoError(t, err)
+	pp, err := poll.NewGovernanceChainCommitteeProtocol(
+		nil,
 		committee,
 		uint64(123456),
 		func(uint64) (time.Time, error) { return time.Now(), nil },
-		5,
-		5,
 		cfg.Chain.PollInitialCandidatesInterval,
-		nil,
-		cfg.Genesis.ProductivityThreshold,
-		cfg.Genesis.KickoutEpochPeriod,
 		cfg.Genesis.KickoutIntensityRate,
-		cfg.Genesis.UnproductiveDelegateMaxCacheSize,
+		slasher,
 	)
 	require.NoError(t, err)
 	require.NoError(t, rp.Register(registry))
